services/chat/requirements.txt
anyio==3.7.0
certifi==2023.5.7
cffi==1.15.1
circuitbreaker==1.4.0
click==8.1.3
cryptography==39.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
oci==2.105.0
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.10
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
uvicorn==0.22.0
